TS by week
Finished making predictions of the TS by week but found a logical problem with how the training and test are divided. So the TS is essentially a moving average, predictions would work for dates after the time period of the data, not the last week of the month for every month. Now going to experiment with collapsing by month so the first 3 weeks in the month can be an input for the last week.