finished TS, without merge
Also  renamed the initial "train" set to be "full" because we later divide that up into train/dev. It was really confusing me and keep that straight (ie before or after cutting out dev, especially when merging). So now, whenever you see the dataset "train" it implies it is only 80% of the original imported dataset because the dev set has been cut out.